Refactored to remove Actor class from models.py
