+++ app/server/game_functions.py
@@ -10,7 +10,7 @@ from sqlalchemy import asc
 from  sqlalchemy.sql.expression import func, select
 
 # Import module models (i.e. Company, Employee, Actor, DNSRecord)
-from app.server.models import db, Company, Employee, Actor, DNSRecord, Team, Users, Roles, GameSession
+from app.server.models import db, Company, Employee, DNSRecord, Team, Users, Roles, GameSession
 from app.server.modules.organization.Company import CompanyShell, EmployeeShell
 from app.server.modules.organization.company_controller import create_company
 from app.server.modules.clock.Clock import Clock
@@ -19,6 +19,7 @@ from app.server.modules.email.email_controller import gen_email
 from app.server.modules.outbound_browsing.browsing_controller import *
 from app.server.modules.outbound_browsing.browsing_controller import browse_random_website
 from app.server.modules.infrastructure.passiveDNS_controller import *
+from app.server.modules.actors.Actor import Actor
 from app.server.utils import *
