+++ app/server/models.py
@@ -81,112 +81,6 @@ class Employee(Base):
             "company_domain": self.company.name
         }
 
-
-class Actor(Base):
-    """ Bad guys doing bad things"""
-    
-    name                = db.Column(db.String(50), nullable=False)
-    effectiveness       = db.Column(db.Integer)
-
-    # makes these long strings and just split them on use
-    domain_themes       = db.Column(db.String(300))
-    sender_themes       = db.Column(db.String(300))
-    subject_themes      = db.Column(db.String(300))
-    file_names          = db.Column(db.String(300))
-    file_extensions     = db.Column(db.String(300))
-    tlds                = db.Column(db.String(300))
-    spoof_email         = db.Column(db.Boolean)
-
-    count_init_passive_dns              = db.Column(db.Integer)
-    count_init_email                    = db.Column(db.Integer)
-    count_init_browsing                 = db.Column(db.Integer)   # >:D
-
-    def __init__(self, name:str, effectiveness:int=50, domain_themes:list=[], sender_themes:list=[], 
-                subject_themes:list=[],  tlds:list=[], spoof_email:bool=False, 
-                count_init_passive_dns:int=100, count_init_email:int=10, count_init_browsing:int=2, 
-                file_names:str="", file_extensions:str="" ):
-
-        self.name = name
-        self.effectiveness      = effectiveness
-        
-        # if any of these values are provided as strings, convert them to lists
-        if isinstance(tlds, list):
-            tlds = " ".join(tlds)
-        
-
-        # we can't have lists in a database, hence the funny business here
-        # take in the list as a space delimited string - then split
-        self.domain_themes              = " ".join(domain_themes + wordGenerator.get_words(10))  # adding random words for entropy
-        self.sender_themes              = " ".join(sender_themes + wordGenerator.get_words(10))
-        self.subject_themes             = " ".join(subject_themes + wordGenerator.get_words(10))
-        self.file_names                 = " ".join(file_names)
-        self.file_extensions            = " ".join(file_extensions)
-        self.tlds                       = tlds or " ".join(['com','net','biz','org','us'])
-        self.spoof_email                = spoof_email
-        self.count_init_browsing        = int(count_init_browsing)
-        self.count_init_email           = int(count_init_email)
-        self.count_init_passive_dns     = int(count_init_passive_dns)
-
-    
-    def get_file_names(self):
-        file_names = self.file_names.split(" ")
-        return [f for f in file_names if f!='']
-
-
-    def get_file_extensions(self):
-        file_extensions = self.file_extensions.split(" ")
-        return [f for f in file_extensions if f!='']
-
-    def get_domain_name(self):
-        """
-        Assemble a domain name using the list of theme words from the Actor object
-        """
-        separators = ["","-","and","or", "with", "on", "in" ]
-        tlds = self.tlds.split(" ")
-        
-        # if actor is default, let's get a larger list of randomised world
-        if self.name == "Default":
-            domain_themes = wordGenerator.get_words(1000)
-        else:
-            domain_themes = self.domain_themes.split(" ")
-
-        words = random.choices(domain_themes, k=random.randint(2,3))
-        
-
-        domain_components = []
-        for loc, word in enumerate(words):
-            domain_components += word 
-            if loc < len(words)-2:
-                domain_components += random.choice(separators)
-
-        domain = "".join(domain_components) + "." + random.choice(tlds)
-
-        return domain
-        
-
-    def get_email_subject(self):
-        """
-        Assemble a subject line using list of theme words from the Actor object
-        """
-        subject_themes = self.subject_themes.split(" ")
-        return fake.sentence(nb_words=7, ext_word_list=subject_themes)
-
-
-    def get_sender_address(self):
-        """Make a list of fake sender addresses"""
-        sender_themes = self.sender_themes.split(" ")
-
-        free_email_service_domains = ['yahoo.com', 'gmail.com', 'aol.com', 'verizon.com', 'yandex.com','hotmail.com','protonmail.com','qq.com']
-        words = random.choices(sender_themes, k=random.randint(2,5))
-        
-        sender_addr = "".join(words) + "@" + random.choice(free_email_service_domains)
-        
-        return sender_addr
-
-    def __repr__(self):
-        return '<Actor %r>' % self.name
-
-
 class DNSRecord(Base):
     """ 
     Belongs to an actor
