+++ app/server/modules/email/email_controller.py
@@ -13,6 +13,7 @@ from app.server.modules.outbound_browsing.browsing_controller import browse_webs
 from app.server.modules.logging.uploadLogs import LogUploader
 from app.server.modules.clock.Clock import Clock
 from app.server.modules.triggers.Trigger import Trigger
+from app.server.modules.actors.Actor import Actor
 from app.server.utils import *
 
 # instantiate faker
